+++ src/Classes/GeneratorInfo.php
@@ -10,4 +10,9 @@ class GeneratorInfo extends BaseModel
 {
     protected string $name = 'Laravel API Factory';
     protected string $version;
+
+    public function getGeneratorTitle(): string
+    {
+        return $this->name . ' v.' . $this->version;
+    }
 }