+++ src/Classes/Generators/TestsGenerator.php
@@ -6,40 +6,41 @@ use Dminustin\ApiFactory\Classes\AbstractClasses\AbstractGenerator;
 use Dminustin\ApiFactory\Classes\ClassAttributes;
 use Dminustin\ApiFactory\Classes\RouterConfig\EndPoint;
 
-class HttpControllerGenerator extends AbstractGenerator
+class TestsGenerator extends AbstractGenerator
 {
-    protected string $name = 'Controller generator';
+    protected string $name = 'Tests generator';
 
     protected function run()
     {
-        $template = $this->loadTemplate('api_factory_controller');
+        $template = $this->loadTemplate('api_factory_test');
 
         /**
          * @var EndPoint $route
          */
         foreach ($this->routes as $route) {
-            $classAttributes = new ClassAttributes(
-                $route->path,
-                $this->config->generatedControllerPathPrefix,
-                $this->config->generatedControllerNSPrefix,
-                $this->config->generatedControllerNSSuffix,
-            );
             $actionAttributes = new ClassAttributes(
                 $route->path,
                 $this->config->generatedActionPathPrefix,
                 $this->config->generatedActionNSPrefix,
                 $this->config->generatedActionNSSuffix,
             );
+            $testAttributes = new ClassAttributes(
+                $route->path,
+                'tests/Unit/ApiFactory/',
+                'Tests\\Unit\\ApiFactory',
+                'ActionTest',
+            );
 
             $this->renderingMap = [
-                'namespace' => $classAttributes->getNameSpace(),
-                'className' => $classAttributes->getShortClassName(),
-                'actionName' => $actionAttributes->getClassName(),
+                'className' => $testAttributes->getShortClassName(),
+                'namespace' => $testAttributes->getNameSpace(),
+                'actionNamespace' => $actionAttributes->getNameSpace(),
+                'actionName' => $actionAttributes->getShortClassName(),
                 'generator' => $this->info->getGeneratorTitle()
             ];
 
             $rendered = $this->render($template);
-            $this->saveFile($classAttributes->getFilePath(), $rendered, $this->config->overrideControllers);
+            $this->saveFile($testAttributes->getFilePath(), $rendered, $this->config->overrideTests);
         }
     }
 }